requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Natural Language Processing
nltk>=3.7
textblob>=0.17.1

# Web Framework (for future backend)
fastapi>=0.68.0
uvicorn[standard]>=0.15.0

# HTTP requests
requests>=2.28.0

# Data validation
pydantic>=1.10.0

# CORS support
fastapi-cors>=0.0.6

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: For advanced NLP features
scikit-learn>=1.1.0

torch>=1.11.0
transformers>=4.20.0
sentencepiece>=0.1.96
Pillow>=9.0.0
tensorflow>=2.12.0
opencv-python
tf-keras
hf_xet